feat: log descriptive cron schedules
