+++ lib/workers/repository/update/branch/status-checks.spec.ts
@@ -7,7 +7,6 @@ import {
   setConfidence,
   setStability,
 } from './status-checks';
-import cronstrue from 'cronstrue';
 describe('workers/repository/update/branch/status-checks', () => {
   describe('setStability', () => {
     let config: StabilityConfig;
@@ -182,22 +181,4 @@ describe('workers/repository/update/branch/status-checks', () => {
       expect(await resolveBranchStatus('somebranch', true, true)).toBe('green');
     });
   });
-  
-  describe('cronstrue', () => {
-    it('should correctly convert "0 22 4 * *" to human-readable format', () => {
-      const result = cronstrue.toString("0 22 4 * *");
-      expect(result).toBe("At 10:00 PM, on day 4 of the month");
-    });
-
-    it('should correctly convert "*/2 * * * *" to human-readable format', () => {
-      const result = cronstrue.toString("*/2 * * * *");
-      expect(result).toBe("Every 2 minutes");
-    });
-
-    it('should correctly convert "0 23 * * *" to human-readable format', () => {
-      const result = cronstrue.toString("0 23 * * *");
-      expect(result).toBe("At 11:00 PM");
-    });
-  });
-
 });
